Add python script to get scoutfs status
The intent of the script is to create a basis for creating
multple layers of commands within the scoutfs status command
that will generate JSON, Tabular, or multiline outputs of
certain file system status.

In this commit we start with just the basic metadata device
path. This again is just to get a basic structure setup to build
upon.

Signed-off-by: Bryant G. Duffy-Ly <bduffyly@versity.com>