Add level1 worm support
Add the first level of worm support.  "level1" adds protection for
regular files by setting an expiration time in the value of a known
scoutfs tagged extended attribute.  The set of protections are meant to
avoid the simplest accidental modification of files while still allowing
some file metadata modifications to support archiving.

This is the first case of extending the inode size and format version so
there is a reasonable amount of code in here that is related to dealing
with combinations of format versions and sizes between the persistent
device and code.

Signed-off-by: Bryant G. Duffy-Ly <bduffyly@versity.com>
[zab@versity.com: fixed bugs, dropped _bits and versions, docs, reworded]
Signed-off-by: Zach Brown <zab@versity.com>