Add fs test functions for mounted paths
We have some fs functions which return info based on the test mount nr
as the test has setup.   This refactofs those a bit to also provide
some of the info when the caller has a path in a given mount.  This will
let tests work with scratch mounts a little more easily.

Signed-off-by: Zach Brown <zab@versity.com>