Fix EOF extent in last block
Currently if there is an extent on the last block the code will only set
EOF on ENOENT. In the case that the last block has an extent it wont go
to the next iteration due to iblock <= last. This then doesnt set the EOF
on the last block in these cases. We want to just allow the loop to keep
looping and rely on if (ext.start > last) to protect us from infinite loop.