Fix truncate for O_DIRECT
In the buffered case page tail zeroing happens automatically.
In the O_DIRECT case it does not so we need to add it in our setattr
path just like EXT2. We want to zero the end of the block that contains
i_size during truncate, so we just call block_truncate_page in
set_inode_size.